requirements.txt
# TensorFlow уже установлен в базовом образе (собран из исходников)
numpy==1.26.4
Pillow==10.3.0
